updated user details screen to use pagedgridview
