+++ lib/features/users/screens/user_details_screen.dart
@@ -6,6 +6,7 @@ import 'package:flutter_mon_loan_tracking/features/loan/bloc/loan_bloc.dart';
 import 'package:flutter_mon_loan_tracking/features/main/bloc/menu_selection_cubit.dart';
 import 'package:flutter_mon_loan_tracking/features/users/bloc/user_bloc.dart';
 import 'package:flutter_mon_loan_tracking/models/civil_status_types.dart';
+import 'package:flutter_mon_loan_tracking/models/loan_schedule.dart';
 import 'package:flutter_mon_loan_tracking/models/menu_item.dart';
 import 'package:flutter_mon_loan_tracking/models/user_type.dart';
 import 'package:flutter_mon_loan_tracking/utils/constants.dart';
@@ -14,6 +15,13 @@ import 'package:flutter_mon_loan_tracking/utils/print_utils.dart';
 import 'package:flutter_mon_loan_tracking/widgets/pdf_generator.dart';
 import 'package:flutter_mon_loan_tracking/widgets/widget_utils.dart';
 import 'package:go_router/go_router.dart';
+import 'package:infinite_scroll_pagination/infinite_scroll_pagination.dart';
+
+part 'user_details_screen_small.dart';
+
+part 'user_details_screen_large.dart';
+
+part 'user_details_screen_utils.dart';
 
 class UserDetailsScreen extends StatefulWidget {
   UserDetailsScreen({super.key});
@@ -37,6 +45,8 @@ class _UserDetailsScreenState extends State<UserDetailsScreen> {
   final passwordController = TextEditingController();
   final confirmPasswordController = TextEditingController();
   final birthDateController = TextEditingController();
+  final _pagingController =
+      PagingController<int, LoanSchedule>(firstPageKey: 0);
 
   @override
   void deactivate() {
@@ -54,6 +64,10 @@ class _UserDetailsScreenState extends State<UserDetailsScreen> {
     context.read<LoanBloc>()
       ..getAllLots()
       ..getAllLoans(clearList: true, clientId: widget.userId);
+
+    // _pagingController.addPageRequestListener((pageKey) {
+    //   context.read<LoanBloc>().getAllLoans();
+    // });
   }
 
   @override
@@ -107,46 +121,59 @@ class _UserDetailsScreenState extends State<UserDetailsScreen> {
     final shouldShowAppBar = widget.isMobile() /*|| widget.isProfile*/;
     printd('size: ${MediaQuery.of(context).size}');
 
-    return BlocListener<UserBloc, UserState>(
-      listener: (context, state) {
-        if (state is UserSuccessState) {
-          if (userBloc.selectedUser != null) {
-            final user = userBloc.selectedUser!;
-            lastNameController.text = user.lastName;
-            firstNameController.text = user.firstName;
-            mobileNumberController.text = user.mobileNumber;
-            emailController.text = user.email;
-            civilStatusController.text = user.civilStatus.value;
-            birthDateController.text = user.birthDate;
-          }
-        } else if (state is UserErrorState) {
-          ScaffoldMessenger.of(context)
-              .showSnackBar(SnackBar(content: Text(state.message)));
-        } else if (state is UserLoadingState) {
-          if (state.isLoading) {
-            showDialog(
-                context: context,
-                barrierDismissible: false,
-                builder: (context) {
-                  return const AlertDialog(
-                    content: SizedBox(
-                      width: 80,
-                      height: 80,
-                      child: CircularProgressIndicator(),
-                    ),
-                  );
-                });
-          } else {
-            try {
-              if (Navigator.of(context, rootNavigator: true).canPop()) {
-                Navigator.of(context, rootNavigator: true).pop();
+    return MultiBlocListener(
+      listeners: [
+        BlocListener<UserBloc, UserState>(
+          listener: (context, state) {
+            if (state is UserSuccessState) {
+              if (userBloc.selectedUser != null) {
+                final user = userBloc.selectedUser!;
+                lastNameController.text = user.lastName;
+                firstNameController.text = user.firstName;
+                mobileNumberController.text = user.mobileNumber;
+                emailController.text = user.email;
+                civilStatusController.text = user.civilStatus.value;
+                birthDateController.text = user.birthDate;
               }
-            } catch (err) {
-              printd(err);
+            } else if (state is UserErrorState) {
+              ScaffoldMessenger.of(context)
+                  .showSnackBar(SnackBar(content: Text(state.message)));
+            } else if (state is UserLoadingState) {
+              if (state.isLoading) {
+                showDialog(
+                    context: context,
+                    barrierDismissible: false,
+                    builder: (context) {
+                      return const AlertDialog(
+                        content: SizedBox(
+                          width: 80,
+                          height: 80,
+                          child: CircularProgressIndicator(),
+                        ),
+                      );
+                    });
+              } else {
+                try {
+                  if (Navigator.of(context, rootNavigator: true).canPop()) {
+                    Navigator.of(context, rootNavigator: true).pop();
+                  }
+                } catch (err) {
+                  printd(err);
+                }
+              }
             }
-          }
-        }
-      },
+          },
+        ),
+        BlocListener<LoanBloc, LoanState>(
+          listener: (context, state) {
+            if (state is LoanDisplaySummaryState) {
+              _pagingController.value = PagingState(
+                itemList: loanBloc.clientLoanSchedules,
+              );
+            }
+          },
+        )
+      ],
       child: Scaffold(
         appBar: !shouldShowAppBar
             ? null
@@ -215,837 +242,35 @@ class _UserDetailsScreenState extends State<UserDetailsScreen> {
                 ),
               ),
         body: shortestSide <= Constants.smallScreenShortestSideBreakPoint
-            ? _buildSmallScreenBody(context: context)
-            : _buildLargeScreenBody(context: context),
+            ? buildSmallScreenBody(
+                context: context,
+                userId: widget.userId,
+                isProfile: widget.isProfile,
+                lastNameController: lastNameController,
+                firstNameController: firstNameController,
+                mobileNumberController: mobileNumberController,
+                emailController: emailController,
+                civilStatusController: civilStatusController,
+                passwordController: passwordController,
+                confirmPasswordController: confirmPasswordController,
+                birthDateController: birthDateController,
+                pagingController: _pagingController,
+              )
+            : buildLargeScreenBody(
+                context: context,
+                userId: widget.userId,
+                isProfile: widget.isProfile,
+                lastNameController: lastNameController,
+                firstNameController: firstNameController,
+                mobileNumberController: mobileNumberController,
+                emailController: emailController,
+                civilStatusController: civilStatusController,
+                passwordController: passwordController,
+                confirmPasswordController: confirmPasswordController,
+                birthDateController: birthDateController,
+                pagingController: _pagingController,
+              ),
       ),
     );
-  }
-
-  Widget _buildSmallScreenBody({required BuildContext context}) {
-    final authenticationBloc = BlocProvider.of<AuthenticationBloc>(context);
-    final loanBloc = BlocProvider.of<LoanBloc>(context);
-
-    final userBloc = BlocProvider.of<UserBloc>(context);
-
-    if (widget.userId != null) {
-      userBloc.selectUser(userId: widget.userId!);
-    }
-
-    final screenSize = MediaQuery.of(context).size;
-    final shortestSide = screenSize.shortestSide;
-    var buttonPadding = const EdgeInsets.all(24);
-
-    if (shortestSide < Constants.largeScreenShortestSideBreakPoint) {
-      buttonPadding = const EdgeInsets.all(16);
-    }
-
-    final width = screenSize.width;
-    final computedWidth = width * 0.88;
-    var appBarHeight = screenSize.height * 0.16;
-    var loginContainerRadius = Constants.defaultRadius;
-    var loginContainerMarginTop = 64.0;
-    var titleTextStyle = Theme.of(context).textTheme.displaySmall;
-    var avatarTextStyle = Theme.of(context).textTheme.titleLarge;
-    var avatarSize = 56.0;
-    var contentPadding = const EdgeInsets.all(58);
-    var appBarBottomPadding = 48.0;
-
-    if (appBarHeight > Constants.maxAppBarHeight) {
-      appBarHeight = Constants.maxAppBarHeight;
-    }
-
-    if (shortestSide < Constants.largeScreenShortestSideBreakPoint) {
-      loginContainerRadius = const Radius.circular(64);
-      loginContainerMarginTop = 32;
-      titleTextStyle = Theme.of(context).textTheme.headlineMedium;
-      avatarTextStyle = Theme.of(context).textTheme.titleSmall;
-      avatarSize = 48;
-      contentPadding = const EdgeInsets.only(
-        left: 32,
-        right: 32,
-        top: 16,
-        bottom: 16,
-      );
-      appBarBottomPadding = 24;
-    }
-    return ListView(
-      children: [
-        const SizedBox(
-          height: 16,
-        ),
-        Padding(
-          padding: EdgeInsets.all(16),
-          child: Column(
-            children: [
-              TextFormField(
-                controller: lastNameController,
-                decoration: const InputDecoration(
-                    label: Text('Last name'), border: OutlineInputBorder()),
-              ),
-              const SizedBox(
-                height: 32,
-              ),
-              TextFormField(
-                controller: firstNameController,
-                decoration: const InputDecoration(
-                    label: Text('First name'), border: OutlineInputBorder()),
-              ),
-              const SizedBox(
-                height: 32,
-              ),
-              Text(
-                'Civil status',
-                style: Theme.of(context).textTheme.titleMedium,
-              ),
-              const SizedBox(
-                height: 8,
-              ),
-              SizedBox(
-                width: double.infinity,
-                child: BlocBuilder<UserBloc, UserState>(
-                  builder: (context, state) {
-                    var dropdownValue = CivilStatus.values.first;
-
-                    if (userBloc.selectedCivilStatus != null) {
-                      dropdownValue = userBloc.selectedCivilStatus!;
-                    }
-
-                    if (state is SelectedCivilStatusState) {
-                      dropdownValue = state.civilStatus;
-                    }
-
-                    return DropdownButton<CivilStatus>(
-                      value: dropdownValue,
-                      items: CivilStatus.values.map((category) {
-                        return DropdownMenuItem<CivilStatus>(
-                          value: category,
-                          child: Text(category.name),
-                        );
-                      }).toList(),
-                      onChanged: (value) => userBloc.selectCivilStatus(
-                        civilStatus: value,
-                      ),
-                    );
-                  },
-                ),
-              ),
-              const SizedBox(
-                height: 32,
-              ),
-              BlocBuilder<UserBloc, UserState>(
-                buildWhen: (previous, current) => current is UserSuccessState,
-                builder: (context, state) {
-                  return TextFormField(
-                    onTap: () {
-                      showDatePicker(
-                        context: context,
-                        initialDate: DateTime.now(),
-                        firstDate:
-                        DateTime.now().subtract(Duration(days: 365 * 100)),
-                        lastDate: DateTime.now(),
-                      ).then((date) {
-                        printd('date is $date');
-                        if (date != null) {
-                          birthDateController.text =
-                              Constants.defaultDateFormat.format(date);
-                          userBloc.selectDate(date: date);
-                        }
-                      });
-                    },
-                    controller: birthDateController,
-                    decoration: const InputDecoration(
-                      label: Text('Birthdate'),
-                      border: OutlineInputBorder(),
-                    ),
-                  );
-                },
-              ),
-              const SizedBox(
-                height: 32,
-              ),
-              TextFormField(
-                controller: mobileNumberController,
-                decoration: const InputDecoration(
-                  label: Text('Mobile number'),
-                  border: OutlineInputBorder(),
-                ),
-                keyboardType:
-                    const TextInputType.numberWithOptions(decimal: true),
-                inputFormatters: [
-                  FilteringTextInputFormatter.allow(RegExp('[0-9.,]'))
-                ],
-              ),
-              const SizedBox(
-                height: 32,
-              ),
-              TextFormField(
-                controller: emailController,
-                decoration: const InputDecoration(
-                  label: Text('Email'),
-                  border: OutlineInputBorder(),
-                ),
-              ),
-              const SizedBox(
-                height: 32,
-              ),
-              Visibility(
-                visible: _shouldShowUpdateButton(),
-                child: Row(
-                  children: [
-                    Expanded(
-                      child: ElevatedButton(
-                        onPressed: () => userBloc.updateUser(
-                          lastName: lastNameController.text,
-                          firstName: firstNameController.text,
-                          birthDate: birthDateController.text,
-                          mobileNumber: mobileNumberController.text,
-                          email: emailController.text,
-                        ),
-                        style: ElevatedButton.styleFrom(
-                            padding: buttonPadding,
-                            backgroundColor:
-                                Theme.of(context).colorScheme.primary),
-                        child: Text(
-                          'Update profile',
-                          style: Theme.of(context)
-                              .textTheme
-                              .titleMedium
-                              ?.apply(color: Colors.white),
-                        ),
-                      ),
-                    ),
-                    SizedBox(
-                      width: 32,
-                    ),
-                    Expanded(
-                      child: ElevatedButton(
-                        onPressed: authenticationBloc.logout,
-                        style: ElevatedButton.styleFrom(
-                            padding: buttonPadding,
-                            backgroundColor:
-                                Theme.of(context).colorScheme.errorContainer),
-                        child: Text(
-                          'Logout',
-                          style: Theme.of(context).textTheme.titleMedium?.apply(
-                                color: Theme.of(context).colorScheme.error,
-                              ),
-                        ),
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-              Visibility(
-                visible: userBloc.getLoggedInUser()?.type == UserType.admin &&
-                    userBloc.getLoggedInUser()?.id != widget.userId,
-                child: Column(
-                  children: [
-                    const SizedBox(
-                      height: 32,
-                    ),
-                    SizedBox(
-                      width: double.infinity,
-                      child: ElevatedButton(
-                        onPressed: authenticationBloc.logout,
-                        style: ElevatedButton.styleFrom(
-                            padding: buttonPadding,
-                            backgroundColor:
-                                Theme.of(context).colorScheme.errorContainer),
-                        child: Text(
-                          'Delete',
-                          style: Theme.of(context).textTheme.titleMedium?.apply(
-                                color: Theme.of(context).colorScheme.error,
-                              ),
-                        ),
-                      ),
-                    )
-                  ],
-                ),
-              ),
-            ],
-          ),
-        ),
-        // const Divider(
-        //   thickness: 1.5,
-        // ),
-        // Padding(
-        //   padding: EdgeInsets.all(16),
-        //   child: Column(
-        //     crossAxisAlignment: CrossAxisAlignment.start,
-        //     children: [
-        //       Row(
-        //         children: [
-        //           Text(
-        //             'Purchased lots',
-        //             style: Theme.of(context).textTheme.titleLarge,
-        //           ),
-        //         ],
-        //       ),
-        //       const SizedBox(
-        //         height: 16,
-        //       ),
-        //       Text('Some purchased lots here ???'),
-        //     ],
-        //   ),
-        // ),
-        const Divider(
-          thickness: 1.5,
-        ),
-        Padding(
-          padding: EdgeInsets.all(16),
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Text(
-                'Loan schedule',
-                style: Theme.of(context).textTheme.titleLarge,
-              ),
-              ElevatedButton(
-                  onPressed: () {
-                    var user = userBloc.selectedUser;
-                    var loan = loanBloc.selectedLoan;
-                    var lot = loanBloc.selectedLot;
-
-                    if (user == null || loan == null || lot == null) {
-                      return;
-                    }
-
-                    PdfGenerator.generatePdf(
-                      user: user,
-                      schedules: loanBloc.clientLoanSchedules,
-                      loan: loan,
-                      lot: lot,
-                    );
-                  },
-                  child: Text('Generate PDF'))
-            ],
-          ),
-        ),
-        const SizedBox(
-          height: 32,
-        ),
-        BlocBuilder<LoanBloc, LoanState>(
-          buildWhen: (previous, current) {
-            return current is LoanSuccessState;
-          },
-          builder: (context, state) {
-            return SingleChildScrollView(
-              scrollDirection: Axis.horizontal,
-              child: DataTable(
-                dataRowHeight: 72,
-                headingRowColor: MaterialStateColor.resolveWith(
-                  (states) => Theme.of(context)
-                      .colorScheme
-                      .secondaryContainer
-                      .withOpacity(0.32),
-                ),
-                columns: [
-                  for (String name
-                      in Constants.user_loan_schedule_table_columns)
-                    DataColumn(
-                        label: Text(
-                      name.toUpperCase(),
-                      style: Theme.of(context).textTheme.titleMedium?.apply(
-                            fontWeightDelta: 3,
-                            color: Theme.of(context).colorScheme.secondary,
-                          ),
-                    ))
-                ],
-                rows: loanBloc.clientLoanSchedules
-                    .map(
-                      (schedule) => DataRow(
-                        cells: [
-                          DataCell(
-                            defaultCellText(
-                              text: Constants.defaultDateFormat.format(
-                                DateTime.fromMillisecondsSinceEpoch(
-                                  schedule.date.toInt(),
-                                ),
-                              ),
-                            ),
-                          ),
-                          DataCell(
-                            Center(
-                              child: defaultCellText(
-                                text: schedule.outstandingBalance.toCurrency(),
-                              ),
-                            ),
-                          ),
-                          DataCell(
-                            Center(
-                              child: defaultCellText(
-                                text: schedule.monthlyAmortization.toCurrency(),
-                              ),
-                            ),
-                          ),
-                          DataCell(
-                            Center(
-                              child: defaultCellText(
-                                text: schedule.principalPayment.toCurrency(),
-                              ),
-                            ),
-                          ),
-                          DataCell(
-                            Center(
-                              child: defaultCellText(
-                                text: schedule.interestPayment.toCurrency(),
-                              ),
-                            ),
-                          ),
-                          DataCell(
-                            Center(
-                              child: defaultCellText(
-                                text: schedule.incidentalFee.toCurrency(),
-                              ),
-                            ),
-                          ),
-                          DataCell(
-                            paymentStatusWidget(
-                              context: context,
-                              schedule: schedule,
-                              loanBloc: loanBloc,
-                              userBloc: userBloc,
-                            ),
-                          ),
-                        ],
-                      ),
-                    )
-                    .toList(),
-              ),
-            );
-          },
-        ),
-        const SizedBox(
-          height: 32,
-        ),
-      ],
-    );
-  }
-
-  Widget _buildLargeScreenBody({required BuildContext context}) {
-    final authenticationBloc = BlocProvider.of<AuthenticationBloc>(context);
-    final loanBloc = BlocProvider.of<LoanBloc>(context);
-
-    final userBloc = BlocProvider.of<UserBloc>(context);
-
-    if (widget.userId != null) {
-      userBloc.selectUser(userId: widget.userId!);
-    }
-
-    final screenSize = MediaQuery.of(context).size;
-    final shortestSide = screenSize.shortestSide;
-    var buttonPadding = const EdgeInsets.all(24);
-
-    if (shortestSide < Constants.largeScreenShortestSideBreakPoint) {
-      buttonPadding = const EdgeInsets.all(16);
-    }
-
-    final width = screenSize.width;
-    final computedWidth = width * 0.88;
-    var appBarHeight = screenSize.height * 0.16;
-    var loginContainerRadius = Constants.defaultRadius;
-    var loginContainerMarginTop = 64.0;
-    var titleTextStyle = Theme.of(context).textTheme.displaySmall;
-    var avatarTextStyle = Theme.of(context).textTheme.titleLarge;
-    var avatarSize = 56.0;
-    var contentPadding = const EdgeInsets.all(58);
-    var appBarBottomPadding = 48.0;
-
-    if (appBarHeight > Constants.maxAppBarHeight) {
-      appBarHeight = Constants.maxAppBarHeight;
-    }
-
-    if (shortestSide < Constants.largeScreenShortestSideBreakPoint) {
-      loginContainerRadius = const Radius.circular(64);
-      loginContainerMarginTop = 32;
-      titleTextStyle = Theme.of(context).textTheme.headlineMedium;
-      avatarTextStyle = Theme.of(context).textTheme.titleSmall;
-      avatarSize = 48;
-      contentPadding = const EdgeInsets.only(
-        left: 32,
-        right: 32,
-        top: 16,
-        bottom: 16,
-      );
-      appBarBottomPadding = 24;
-    }
-
-    return ListView(
-      children: [
-        const SizedBox(
-          height: 16,
-        ),
-        TextFormField(
-          controller: lastNameController,
-          decoration: const InputDecoration(
-              label: Text('Last name'), border: OutlineInputBorder()),
-        ),
-        const SizedBox(
-          height: 32,
-        ),
-        TextFormField(
-          controller: firstNameController,
-          decoration: const InputDecoration(
-              label: Text('First name'), border: OutlineInputBorder()),
-        ),
-        const SizedBox(
-          height: 32,
-        ),
-        Row(
-          children: [
-            Column(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                Text(
-                  'Civil status',
-                  style: Theme.of(context).textTheme.titleMedium,
-                ),
-                const SizedBox(
-                  height: 8,
-                ),
-                BlocBuilder<UserBloc, UserState>(
-                  builder: (context, state) {
-                    var dropdownValue = CivilStatus.values.first;
-
-                    if (state is SelectedCivilStatusState) {
-                      dropdownValue = state.civilStatus;
-                    }
-
-                    return DropdownButton<CivilStatus>(
-                      value: dropdownValue,
-                      items: CivilStatus.values.map((category) {
-                        return DropdownMenuItem<CivilStatus>(
-                          value: category,
-                          child: Text(category.value),
-                        );
-                      }).toList(),
-                      onChanged: (value) => userBloc.selectCivilStatus(
-                        civilStatus: value,
-                      ),
-                    );
-                  },
-                ),
-              ],
-            ),
-            const SizedBox(
-              width: 32,
-            ),
-            Expanded(
-              child: BlocBuilder<UserBloc, UserState>(
-                buildWhen: (previous, current) => current is UserSuccessState,
-                builder: (context, state) {
-                  return TextFormField(
-                    onTap: () {
-                      showDatePicker(
-                        context: context,
-                        initialDate: DateTime.now(),
-                        firstDate:
-                        DateTime.now().subtract(Duration(days: 365 * 100)),
-                        lastDate: DateTime.now(),
-                      ).then((date) {
-                        printd('date is $date');
-                        if (date != null) {
-                          birthDateController.text =
-                              Constants.defaultDateFormat.format(date);
-                          userBloc.selectDate(date: date);
-                        }
-                      });
-                    },
-                    controller: birthDateController,
-                    decoration: const InputDecoration(
-                      label: Text('Birthdate'),
-                      border: OutlineInputBorder(),
-                    ),
-                  );
-                },
-              ),
-            ),
-            const SizedBox(
-              width: 32,
-            ),
-            Expanded(
-              child: TextFormField(
-                controller: mobileNumberController,
-                decoration: const InputDecoration(
-                  label: Text('Mobile number'),
-                  border: OutlineInputBorder(),
-                ),
-                keyboardType:
-                    const TextInputType.numberWithOptions(decimal: true),
-                inputFormatters: [
-                  FilteringTextInputFormatter.allow(RegExp('[0-9.,]'))
-                ],
-              ),
-            ),
-            const SizedBox(
-              width: 32,
-            ),
-            Expanded(
-              child: TextFormField(
-                controller: emailController,
-                decoration: const InputDecoration(
-                  label: Text('Email'),
-                  border: OutlineInputBorder(),
-                ),
-              ),
-            ),
-          ],
-        ),
-        const SizedBox(
-          height: 32,
-        ),
-        Row(
-          children: [
-            Visibility(
-              visible: _shouldShowUpdateButton(),
-              child: Expanded(
-                child: ElevatedButton(
-                  onPressed: () => userBloc.updateUser(
-                    lastName: lastNameController.text,
-                    firstName: firstNameController.text,
-                    birthDate: birthDateController.text,
-                    mobileNumber: mobileNumberController.text,
-                    email: emailController.text,
-                  ),
-                  style: ElevatedButton.styleFrom(
-                      padding: buttonPadding,
-                      backgroundColor: Theme.of(context).colorScheme.primary),
-                  child: Text(
-                    'Update profile',
-                    style: Theme.of(context)
-                        .textTheme
-                        .titleMedium
-                        ?.apply(color: Colors.white),
-                  ),
-                ),
-              ),
-            ),
-            SizedBox(
-              width: 32,
-            ),
-            Visibility(
-              visible: widget.isProfile,
-                child: Expanded(
-                  child: ElevatedButton(
-                    onPressed: authenticationBloc.logout,
-                    style: ElevatedButton.styleFrom(
-                        padding: buttonPadding,
-                        backgroundColor: Theme.of(context).colorScheme.error),
-                    child: Text(
-                      'Logout',
-                      style: Theme.of(context)
-                          .textTheme
-                          .titleMedium
-                          ?.apply(color: Colors.white),
-                    ),
-                  ),
-                )
-            ),
-          ],
-        ),
-        Visibility(
-          visible: userBloc.getLoggedInUser()?.type == UserType.admin &&
-              userBloc.getLoggedInUser()?.id != widget.userId,
-          child: Column(
-            children: [
-              const SizedBox(
-                height: 32,
-              ),
-              SizedBox(
-                width: double.infinity,
-                child: ElevatedButton(
-                  onPressed: authenticationBloc.logout,
-                  style: ElevatedButton.styleFrom(
-                      padding: buttonPadding,
-                      backgroundColor:
-                          Theme.of(context).colorScheme.errorContainer),
-                  child: Text(
-                    'Delete',
-                    style: Theme.of(context).textTheme.titleMedium?.apply(
-                          color: Theme.of(context).colorScheme.error,
-                        ),
-                  ),
-                ),
-              )
-            ],
-          ),
-        ),
-        const SizedBox(
-          height: 32,
-        ),
-        // const Divider(
-        //   thickness: 1.5,
-        // ),
-        // const SizedBox(
-        //   height: 16,
-        // ),
-        // Row(
-        //   children: [
-        //     Text(
-        //       'Purchased lots',
-        //       style: Theme.of(context).textTheme.titleLarge,
-        //     ),
-        //   ],
-        // ),
-        // const SizedBox(
-        //   height: 16,
-        // ),
-        // Text('Some purchased lots here ???'),
-        // const SizedBox(
-        //   height: 32,
-        // ),
-        const Divider(
-          thickness: 1.5,
-        ),
-        const SizedBox(
-          height: 32,
-        ),
-        Row(
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          children: [
-            Text(
-              'Loan schedule',
-              style: Theme.of(context).textTheme.titleLarge,
-            ),
-            ElevatedButton(
-                onPressed: () {
-                  var user = userBloc.selectedUser;
-                  var loan = loanBloc.selectedLoan;
-                  var lot = loanBloc.selectedLot;
-
-                  if (user == null || loan == null || lot == null) {
-                    return;
-                  }
-
-                  PdfGenerator.generatePdf(
-                    user: user,
-                    schedules: loanBloc.clientLoanSchedules,
-                    loan: loan,
-                    lot: lot,
-                  );
-                },
-                child: Text(
-                  'Generate PDF',
-                  style: Theme.of(context).textTheme.labelLarge,
-                ))
-          ],
-        ),
-        const SizedBox(
-          height: 32,
-        ),
-        BlocBuilder<LoanBloc, LoanState>(
-          buildWhen: (previous, current) {
-            return current is LoanSuccessState;
-          },
-          builder: (context, state) {
-            return SingleChildScrollView(
-              scrollDirection: Axis.horizontal,
-              child: DataTable(
-                dataRowHeight: 72,
-                headingRowColor: MaterialStateColor.resolveWith(
-                      (states) => Theme.of(context)
-                      .colorScheme
-                      .secondaryContainer
-                      .withOpacity(0.32),
-                ),
-                columns: [
-                  for (String name in Constants.user_loan_schedule_table_columns)
-                    DataColumn(
-                        label: Text(
-                          name.toUpperCase(),
-                          style: Theme.of(context).textTheme.titleMedium?.apply(
-                            fontWeightDelta: 3,
-                            color: Theme.of(context).colorScheme.secondary,
-                          ),
-                        ))
-                ],
-                rows: loanBloc.clientLoanSchedules
-                    .map(
-                      (schedule) => DataRow(
-                    cells: [
-                      DataCell(
-                        defaultCellText(
-                          text: Constants.defaultDateFormat.format(
-                            DateTime.fromMillisecondsSinceEpoch(
-                              schedule.date.toInt(),
-                            ),
-                          ),
-                        ),
-                      ),
-                      DataCell(
-                        Center(
-                          child: defaultCellText(
-                            text: schedule.outstandingBalance.toCurrency(),
-                          ),
-                        ),
-                      ),
-                      DataCell(
-                        Center(
-                          child: defaultCellText(
-                            text: schedule.monthlyAmortization.toCurrency(),
-                          ),
-                        ),
-                      ),
-                      DataCell(
-                        Center(
-                          child: defaultCellText(
-                            text: schedule.principalPayment.toCurrency(),
-                          ),
-                        ),
-                      ),
-                      DataCell(
-                        Center(
-                          child: defaultCellText(
-                            text: schedule.interestPayment.toCurrency(),
-                          ),
-                        ),
-                      ),
-                      DataCell(
-                        Center(
-                          child: defaultCellText(
-                            text: schedule.incidentalFee.toCurrency(),
-                          ),
-                        ),
-                      ),
-                      DataCell(
-                        paymentStatusWidget(
-                          context: context,
-                          schedule: schedule,
-                          loanBloc: loanBloc,
-                          userBloc: userBloc,
-                        ),
-                      ),
-                    ],
-                  ),
-                )
-                    .toList(),
-              ),
-            );
-          },
-        ),
-        const SizedBox(
-          height: 32,
-        ),
-      ],
-    );
-  }
-
-  /// show update button if user is
-  /// detail is the current user or
-  /// if user is subadmin
-  bool _shouldShowUpdateButton() {
-    final userBloc = BlocProvider.of<UserBloc>(context);
-
-    if(userBloc.getLoggedInUser() != null &&
-        [UserType.admin, UserType.subAdmin]
-            .contains(userBloc.getLoggedInUser()!.type)) {
-      return true;
-    }
-
-    if (widget.userId != null && userBloc.getLoggedInUser() != null) {
-      return widget.userId == userBloc.getLoggedInUser()!.id;
-    }
-
-    return false;
   }
 }
